Moves to tray based application
The application now runs as a 'tray' application, with an
icon in the system tray. Have also added ability to enable/disable
showing of the 'break' screen on given monitors; i.e., if you have
dual monitor setup, can enable windows to only display on primary
or secondary monitor. There is also a countdown timer in the tray
dialog that informs how much time is remaining until next break.

v0.1.0